requirements.txt
Django>=1.9
python-social-auth>=0.2.3
pycrest
python-dateutil
